JVM: extract special builtins mapping from MethodSignatureMapper
 #KT-84562
