+++ core/descriptors.jvm/src/org/jetbrains/kotlin/load/java/specialBuiltinMembers.kt
@@ -116,23 +116,6 @@ fun <T : CallableMemberDescriptor> T.getOverriddenSpecialBuiltin(): T? {
     } as T?
 }
 
-// The subtle difference between getOverriddenBuiltinReflectingJvmDescriptor and getOverriddenSpecialBuiltin
-// is that first one return descriptor reflecting JVM signature (JVM descriptor)
-// E.g. it returns `contains(e: E): Boolean` instead of `contains(e: String): Boolean` for implementation of Collection<String>.contains
-// Implementation differs by getting 'original' for collection methods with erased value parameters
-// Also it ignores Collection<String>.containsAll overrides because they have the same JVM descriptor
-@Suppress("UNCHECKED_CAST")
-@K1Deprecation
-fun <T : CallableMemberDescriptor> T.getOverriddenBuiltinReflectingJvmDescriptor(): T? {
-    getOverriddenBuiltinWithDifferentJvmName()?.let { return it }
-
-    if (!name.sameAsBuiltinMethodWithErasedValueParameters) return null
-
-    return firstOverridden {
-        KotlinBuiltIns.isBuiltIn(it) && it.getSpecialSignatureInfo()?.isObjectReplacedWithTypeParameter ?: false
-    }?.original as T?
-}
-
 @K1Deprecation
 fun getJvmMethodNameIfSpecial(callableMemberDescriptor: CallableMemberDescriptor): String? {
     val overriddenBuiltin = getOverriddenBuiltinThatAffectsJvmName(callableMemberDescriptor)?.propertyIfAccessor
